moved functions in embed_tweets to utils script for use elsewhere
